refactor(suite-native): dedupe standalone wrap and unwrap form screens
